Updating for Google Fonts
I've created the build script for Hahmlet that creates the TTF static instances and also generates the font as a variable font. All versions are then autohinted.

As part of this work, I have made a number of changes to the source file.

- fixed glyph widths of several glyphs that were incorrect: ㆈ ㆉ ㆊ ㆋ ㆌ 땎
- adjusted empty glyph widths for consistency
- removed "_" anchors from PÞ‼№ㅸㆁ. which were problematic
- rebuilt colon as removal of anchors in period caused problems
- adjusted position of 7 in ⅞
- MVAR is not well supported currently and can lead to various problems :(, so adjustments disabled for Hahmlet Variable
- cleaned up metadata and glyphs file

One thing that isn't currently implemented is PANOSE, which I'll look into next, but please take a look at the font files and let me know if you run into any issues.