Now updated with Panose values in the static instances
ufo2ft can't generate the panose values from the Glyphs customParameters, and GlyphsLib doesn't insert the data into openTypeOS2Panose, so nothing was coming through. So I've updated the code to read it manually.

No panose values are currently present in the variable font. We can discuss that if you'd like :)